Improved timer management & logging
No longer directly using the timer to determine when to refresh for new grades. Instead, the timer interval is now fixed at 1 second and the event initializes the sessions for the saved users with a random int for each user that will determine when the app will start monitoring grades for this user,.

This means that the users' grades will refresh at different times, and thus slightly reducing network load and fixing the issue where if an error occurred for 1 user it would also change the timer interval of the current user (since they previously used the same timer, which is now not directly the case as even though they use the same timer, they will be triggered at different times).